deploy th cartmicroservice and modify the sign service

+++ backend/MicroService/CartMicroService/src/main/java/com/sjtusummerproject/cartmicroservice/DataModel/Domain/CartEntity.java
@@ -17,7 +17,7 @@ public class CartEntity implements Serializable{
     Long userId;
     Long ticketId;
     double price;
-    int number;
+    Long number;
     String image;
     String title;
     String date;
@@ -72,11 +72,11 @@ public class CartEntity implements Serializable{
         this.price = price;
     }
 
-    public int getNumber() {
+    public Long getNumber() {
         return number;
     }
 
-    public void setNumber(int number) {
+    public void setNumber(Long number) {
         this.number = number;
     }
